fix silly mistakes, tests should run now
